+++ src/bos/server/controllers/v2/components.py
@@ -42,7 +42,7 @@ SESSIONS_DB = dbutils.get_wrapper(db='sessions')
 
 @tenant_error_handler
 @dbutils.redis_error_handler
-def get_v2_components(ids="",
+def get_v2_components(ids=None,
                       enabled=None,
                       session=None,
                       staged_session=None,
@@ -58,7 +58,7 @@ def get_v2_components(ids="",
         "GET /v2/components invoked get_v2_components with ids=%s enabled=%s session=%s "
         "staged_session=%s phase=%s status=%s start_after_id=%s page_size=%d", ids,
         enabled, session, staged_session, phase, status, start_after_id, page_size)
-    if ids:
+    if ids is not None:
         try:
             id_list = ids.split(',')
         except Exception as err:
@@ -68,7 +68,7 @@ def get_v2_components(ids="",
                                      detail=str(err))
     else:
         id_list = None
-    tenant = get_tenant_from_header()
+    tenant = get_tenant_from_header() or None
     LOGGER.debug("GET /v2/components for tenant=%s with %d IDs specified",
                  tenant, len(id_list) if id_list else 0)
     response = get_v2_components_data(id_list=id_list,