+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added basic paging ability for `GET` requests for `components`.
 
+### Fixed
+- Fixed bug causing no components to be listed when no tenant specified.
+
 ## [2.31.1] - 2024-12-18
 ### Fixed
 - When renaming session templates during migration, use correct database key to store renamed template.